final reorgnaization and push of the maker classes
