+++ SimCLR/Blocks.py
@@ -1,6 +1,6 @@
 import torch.nn as nn
-from resnet.Tools import activation_func
-from resnet.Components import Conv2dPadded
+from SimCLR.Tools import activation_func
+from SimCLR.Components import Conv2dPadded
 
 from functools import partial
