+++ SimCLR/Components.py
