+++ SimCLR/__init__.py
@@ -1,7 +1,13 @@
-import SimCLR.Models
-import SimCLR.Loss
+import SimCLR.Components
 import SimCLR.Augment
+import SimCLR.Blocks
+import SimCLR.Layers
+import SimCLR.Models
+import SimCLR.Makers
+import SimCLR.Tools
 
 __all__ = [
-  SimCLR.Models, SimCLR.Loss, SimCLR.Augment
+  SimCLR.Tools, SimCLR.Blocks, SimCLR.Models,
+  SimCLR.Components, SimCLR.Layers,
+  SimCLR.Makers, SimCLR.Loss, SimCLR.Augment
 ]
