--- SimCLR/resnet/__init__.py
@@ -1,10 +1,0 @@
-import resnet.Blocks
-import resnet.Components
-import resnet.Layers
-import resnet.Models
-import resnet.Tools
-
-__all__ = [
-  resnet.Tools, resnet.Blocks, resnet.Models,
-  resnet.Components, resnet.Layers
-]
